services/collector/requirements.txt
telethon
sqlalchemy
psycopg2-binary
python-dotenv
pyyaml 